// src/api-chain/model/docTypeIdsWithRequired.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * coffee-be
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */





export interface DocTypeIdsWithRequired { 
    processingEvidenceTypeId: string;
    required?: boolean;
    requiredOnQuote?: boolean;
    requiredOneOfGroupIdForQuote?: string;
}

/**
 * Namespace for property- and property-value-enumerations of DocTypeIdsWithRequired.
 */
export namespace DocTypeIdsWithRequired {
    /**
     * All properties of DocTypeIdsWithRequired.
     */
    export enum Properties {
        processingEvidenceTypeId = 'processingEvidenceTypeId',
        required = 'required',
        requiredOnQuote = 'requiredOnQuote',
        requiredOneOfGroupIdForQuote = 'requiredOneOfGroupIdForQuote'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'DocTypeIdsWithRequired',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'processingEvidenceTypeId',
                    classname: 'DocTypeIdsWithRequired',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'required',
                    classname: 'DocTypeIdsWithRequired',
                    dataType: 'boolean',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'requiredOnQuote',
                    classname: 'DocTypeIdsWithRequired',
                    dataType: 'boolean',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'requiredOneOfGroupIdForQuote',
                    classname: 'DocTypeIdsWithRequired',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                processingEvidenceTypeId: [
                        ['required'],
                ],
                required: [
                ],
                requiredOnQuote: [
                ],
                requiredOneOfGroupIdForQuote: [
                ],
            }
        }
    }

  // export const DocTypeIdsWithRequiredValidationScheme = {
  //     validators: [],
  //     fields: {
  //               processingEvidenceTypeId: {
  //                   validators: []
  //               },
  //               required: {
  //                   validators: []
  //               },
  //               requiredOnQuote: {
  //                   validators: []
  //               },
  //               requiredOneOfGroupIdForQuote: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<DocTypeIdsWithRequired>;


}